multi-agent/requirements.txt
# Multi-Agent System Requirements

# API Clients
openai>=2.3.0
anthropic>=0.69.0
google-generativeai>=0.8.5

# Configuration
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0

# Optional: For advanced features
# tiktoken>=0.5.0  # Token counting
# rich>=13.0.0     # Better terminal output
# pandas>=2.0.0    # Data analysis
